fix test and codeclimate
