+++ pkg/controller/secret/stringsecret_controller_test.go
@@ -211,6 +211,22 @@ func TestControllerGenerateSecretSingleFieldMixed(t *testing.T) {
 	require.True(t, errors.IsNotFound(err), "Secret was not deleted upon cr deletion")
 }
 
+func TestNewSecret(t *testing.T) {
+	owner := v1alpha1.StringSecret{}
+	owner.Name = "testSecret"
+	owner.Namespace = "testns"
+	owner.Labels = map[string]string{"test": "test"}
+
+	data := map[string][]byte{"test": []byte("blah")}
+
+	target, err := crd.NewSecret(&owner, data, "opaque")
+	require.NoError(t, err)
+	require.Equal(t, "testSecret", target.Name)
+	require.Equal(t, "testns", target.Namespace)
+	require.Equal(t, map[string]string{"test": "test"}, target.Labels)
+
+}
+
 func TestControllerGenerateSecretMultipleFields(t *testing.T) {
 	testSpec := v1alpha1.StringSecretSpec{
 		Type: string(corev1.SecretTypeOpaque),
